pyproject.toml
[project]
name = "side-p"
version = "0.1.0"
description = ""
authors = [
    {name = "MohitSV-ept",email = "mohit.sv@ept.ai"}
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pandas = ">=2.2.3,<3.0.0"
numpy = ">=1.22.4,<2.0.0"  # Updated to match LangChain requirements
pinecone = "^5.4.2"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
langchain-core = "^0.3.33"
llama-index-core = "^0.12.16.post1"
llama-index-utils-workflow = "^0.3.0"
llama-index = "^0.12.16"
langchain-openai = "^0.3.4"
langchain-experimental = "^0.3.4"
tabulate = "^0.9.0"
faiss-cpu = "^1.10.0"
tavily-python = "^0.5.1"


